feat : 재고 처리 기능 추가 및 리팩토링
Product 엔티티 수정
- 재고 처리 메서드 추가(주문 상태에 따른 증감)

주문 조회, 수정, 삭제 시 로그인 필요

totalPrice 타입 변경 Integer -> Long

주석 추가

+++ src/main/java/com/ecommercebackoffice/order/dto/ReadOneOrderResponseDto.java
@@ -13,14 +13,14 @@ public class ReadOneOrderResponseDto {
     private final String customerEmail;
     private final String productName;
     private final Integer quantity;
-    private final Integer totalPrice;
+    private final Long totalPrice;
     private final OrderStatus status;
     private final LocalDateTime createdAt;
     private final String createdByName;
     private final String createdByEmail;
     private final AdminRole createdByRole;
 
-    public ReadOneOrderResponseDto(String orderNumber, String customerName, String customerEmail, String productName, Integer quantity, Integer totalPrice, OrderStatus status, LocalDateTime createdAt, String createdByName, String createdByEmail, AdminRole createdByRole) {
+    public ReadOneOrderResponseDto(String orderNumber, String customerName, String customerEmail, String productName, Integer quantity, Long totalPrice, OrderStatus status, LocalDateTime createdAt, String createdByName, String createdByEmail, AdminRole createdByRole) {
         this.orderNumber = orderNumber;
         this.customerName = customerName;
         this.customerEmail = customerEmail;
